+++ test/bydi-report-test.el
@@ -9,6 +9,7 @@
 (require 'bydi-report)
 
 (defvar coverage-file (ert-resource-file "coverage.txt"))
+(defvar mock-coverage-file (ert-resource-file "mock-coverage.txt"))
 
 (ert-deftest bydi-report--report ()
   (bydi (message)
@@ -23,7 +24,8 @@
 
 (ert-deftest bydi-undercover-setup ()
   (bydi (undercover--setup
-         (:mock getenv :with ignore))
+         (:mock getenv :with ignore)
+         bydi-report--undercover-result)
 
     (bydi-undercover-setup (list "bydi.el"))
 
@@ -51,6 +53,17 @@
                                                (:report-file "./coverage/.resultset.json")
                                                (:send-report nil))))))
 
+(ert-deftest bydi-report--undercover-result ()
+  (let ((bydi-report--text-file mock-coverage-file))
+
+    (ert-with-message-capture messages
+
+      (bydi ((:mock bydi-report--consume-undercover-report :return '(10 3 2 1)))
+        (bydi-report--undercover-result))
+
+      (should (string=
+               "COVERAGE\n\nAverage : Percent 10% [Relevant: 3 Covered: 2 Missed: 1]\n\n"
+               messages)))))
 
 (ert-deftest bydi-ert-runner-setup ()
   (bydi (add-hook)